moved out api stuff to separate file
